backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
filelock
httpx 
python-multipart
pandas 
sqlalchemy 
psycopg2-binary 
python-dotenv 
openpyxl